+++ schema-forwarder/internal/transformer/transformer.go
@@ -1,4 +1,4 @@
-package schematransformer
+package transformer
 
 import (
 	"context"
@@ -27,14 +27,14 @@ type Transformer interface {
 
 // New returns a new instance of Schema Transformer
 func New(backendConfig backendconfig.BackendConfig, config *config.Config) Transformer {
-	return &SchemaTransformer{
+	return &transformer{
 		backendConfig:        backendConfig,
 		captureNilAsUnknowns: config.GetBool("EventSchemas.captureUnknowns", false),
 	}
 }
 
 // Start starts the schema transformer
-func (st *SchemaTransformer) Start() {
+func (st *transformer) Start() {
 	ctx, cancel := context.WithCancel(context.Background())
 	st.cancel = cancel
 	st.g, ctx = errgroup.WithContext(ctx)
@@ -56,13 +56,13 @@ func (st *SchemaTransformer) Start() {
 }
 
 // Stop stops the schema transformer
-func (st *SchemaTransformer) Stop() {
+func (st *transformer) Stop() {
 	st.cancel()
 	_ = st.g.Wait()
 }
 
 // Transform transforms the job into a schema message and returns the schema message along with write key
-func (st *SchemaTransformer) Transform(job *jobsdb.JobT) (*proto.EventSchemaMessage, string, error) {
+func (st *transformer) Transform(job *jobsdb.JobT) (*proto.EventSchemaMessage, string, error) {
 	var eventPayload map[string]interface{}
 	if err := json.Unmarshal(job.EventPayload, &eventPayload); err != nil {
 		return nil, "", err
@@ -80,7 +80,7 @@ func (st *SchemaTransformer) Transform(job *jobsdb.JobT) (*proto.EventSchemaMess
 }
 
 // getSchemaKeyFromJob returns the schema key from the job based on the event type and event identifier
-func (st *SchemaTransformer) getSchemaKeyFromJob(eventPayload map[string]interface{}, writeKey string) *proto.EventSchemaKey {
+func (st *transformer) getSchemaKeyFromJob(eventPayload map[string]interface{}, writeKey string) *proto.EventSchemaKey {
 	eventType := st.getEventType(eventPayload)
 	return &proto.EventSchemaKey{
 		WriteKey:        writeKey,
@@ -89,7 +89,7 @@ func (st *SchemaTransformer) getSchemaKeyFromJob(eventPayload map[string]interfa
 	}
 }
 
-func (st *SchemaTransformer) backendConfigSubscriber(ctx context.Context, loopFn func()) {
+func (st *transformer) backendConfigSubscriber(ctx context.Context, loopFn func()) {
 	ch := st.backendConfig.Subscribe(ctx, backendconfig.TopicProcessConfig)
 	for data := range ch {
 		configData := data.Data.(map[string]backendconfig.ConfigT)
@@ -111,7 +111,7 @@ func (st *SchemaTransformer) backendConfigSubscriber(ctx context.Context, loopFn
 }
 
 // getEventType returns the event type from the event
-func (st *SchemaTransformer) getEventType(event map[string]interface{}) string {
+func (st *transformer) getEventType(event map[string]interface{}) string {
 	eventType, ok := event["type"].(string)
 	if !ok {
 		return ""
@@ -120,7 +120,7 @@ func (st *SchemaTransformer) getEventType(event map[string]interface{}) string {
 }
 
 // getEventIdentifier returns the event identifier from the event
-func (st *SchemaTransformer) getEventIdentifier(event map[string]interface{}, eventType string) string {
+func (st *transformer) getEventIdentifier(event map[string]interface{}, eventType string) string {
 	eventIdentifier := ""
 	if eventType == "track" {
 		eventIdentifier, ok := event["event"].(string)
@@ -133,7 +133,7 @@ func (st *SchemaTransformer) getEventIdentifier(event map[string]interface{}, ev
 }
 
 // getSchemaMessage returns the schema message from the event by flattening the event and getting the schema
-func (st *SchemaTransformer) getSchemaMessage(key *proto.EventSchemaKey, event map[string]interface{}, sample json.RawMessage, workspaceId string, observedAt time.Time) (*proto.EventSchemaMessage, error) {
+func (st *transformer) getSchemaMessage(key *proto.EventSchemaKey, event map[string]interface{}, sample json.RawMessage, workspaceId string, observedAt time.Time) (*proto.EventSchemaMessage, error) {
 	flattenedEvent, err := st.flattenEvent(event)
 	if err != nil {
 		return nil, err
@@ -152,7 +152,7 @@ func (st *SchemaTransformer) getSchemaMessage(key *proto.EventSchemaKey, event m
 }
 
 // getSchema returns the schema from the flattened event
-func (st *SchemaTransformer) getSchema(flattenedEvent map[string]interface{}) map[string]string {
+func (st *transformer) getSchema(flattenedEvent map[string]interface{}) map[string]string {
 	schema := make(map[string]string)
 	for k, v := range flattenedEvent {
 		reflectType := reflect.TypeOf(v)
@@ -166,7 +166,7 @@ func (st *SchemaTransformer) getSchema(flattenedEvent map[string]interface{}) ma
 }
 
 // flattenEvent flattens the event
-func (st *SchemaTransformer) flattenEvent(event map[string]interface{}) (map[string]interface{}, error) {
+func (st *transformer) flattenEvent(event map[string]interface{}) (map[string]interface{}, error) {
 	flattenedEvent, err := flatten.Flatten(event, "", flatten.DotStyle)
 	if err != nil {
 		return nil, err
@@ -175,14 +175,14 @@ func (st *SchemaTransformer) flattenEvent(event map[string]interface{}) (map[str
 }
 
 // disablePIIReporting returns whether PII reporting is disabled for the write key
-func (st *SchemaTransformer) disablePIIReporting(writeKey string) bool {
+func (st *transformer) disablePIIReporting(writeKey string) bool {
 	st.mu.RLock()
 	defer st.mu.RUnlock()
 	return st.newPIIReportingSettings[writeKey]
 }
 
 // getWriteKeyFromParams returns the write key from the job parameters
-func (st *SchemaTransformer) getWriteKeyFromParams(parameters json.RawMessage) string {
+func (st *transformer) getWriteKeyFromParams(parameters json.RawMessage) string {
 	sourceId := gjson.GetBytes(parameters, "source_id").Str
 	if sourceId == "" {
 		return sourceId
